Upgrade acceptance test

+++ contracts/library/Equipment.sol
@@ -8,7 +8,7 @@ interface IRarityEquipment {
   function equip(uint summoner, uint slot_type, address mint, address codex, uint token) external;
   function unequip(uint summoner, uint slot_type) external;
   function snapshot(uint token, uint summoner) external;
-  function snapshots(address encounter, uint token, uint summoner, uint slot_type) external view returns (Equipment.Slot memory);
+  function snapshots(address encounter, uint token, uint summoner, uint8 slot_type) external view returns (Equipment.Slot memory);
 }
 
 library Equipment {
